fix(@angular-devkit/schematics): VirtualDirEntry#subdirs & proper testing
- the getter `subdirs` in class `VirtualDirEntry` was wrongly implemented,
returning irrelevant data. Make some code common between `subdirs`
and `subfiles` for clarity and DRY respect.
- testing was not done properly either, since a `FileSystemTree` instance was used
(creating corresponding `FileSystemDirEntry` instances) in a test for `VirtualTree`
(and its corresponding model `VirtualDirEntry`). Despite the inheritance,
it wouldn't test it properly. Therefore copy this test as is to the proper file
`filesystem_spec.ts` and slightly adapt it in `virtual_spec.ts` to use the proper
implementations instead.